stream/sample_drilling_stream_app/requirements.txt
corva-sdk==0.0.18
